requirements.txt
# Dependencies for Adzuna Job Scraper

# HTTP requests
requests>=2.28.0

# PostgreSQL
psycopg2-binary>=2.9.0

# Airflow (pour développement local si besoin)
apache-airflow>=2.10.0
apache-airflow-providers-postgres>=5.11.0

# Data manipulation and analysis
numpy>=1.21.0,<2.0.0
pandas>=1.5.0

# Jupyter for analysis
jupyter>=1.0.0
ipykernel>=6.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
black>=22.0.0
flake8>=5.0.0
mypy>=0.990
